requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-bootstrap-pagination==1.7.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.6
pytz==2020.4
requests==2.25.0
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
xmltodict==0.12.0
